requirements.txt
streamlit
pandas==2.3.1
openpyxl==3.1.5
python-dotenv==1.1.1
google-generativeai==0.8.5
langgraph==0.2.66
langchain-core==0.3.63
